Avoid running haxelib through the system shell
Trying to stop Haxe from accessing the system shell / command interpreter in order to get path info from haxelib.

`open_process_args_full` is only available with ocaml 4.08, but the minimum version used to build Haxe is 4.02. I've copied and adapted the code from unix.ml to work here. That change is annotated with an "ocaml 4.08" comment.

`find_program_name` is a pretty simplified way of searching in cwd/the path for the named program. It doesn't search exactly like a shell/interpreter would, but it should do what we need it to in most cases.