Fix recorded method type for js stack traces on v8
The original code seems to expect getFunctionName to always return both
the type and the function in the form <type>.<function>, but this isn't
the case for constructors and instance methods (Chrome 126).

Haxe function calls can be divided into four groups:
- Function calls on abstract types (<type>.<function>)
- Static function calls            (<type>.<function>)
- Constructors                     (<type>)
- Function calls on instances      (<function>)

Further details are below about what each generates and how this can
be fixed.

Open Question: how should constructors be handled? With this change,
constructors will appear as Method(TypeName,"new"). How do other
platforms handle constructors in stack traces?

-- Abstracts --

Abstracts are generated as JS objects:

  var haxe_CallStack = {};
  haxe_CallStack.callStack = function() {...};

And function calls on abstract like this:

  haxe_CallStack.callStack()

When this function appears in a v8 stack trace, these
properties are present:

  type: Object
  function: haxe_CallStack.callStack
  method: callStack

This works as expected with the previous code.

-- Classes --

Classes are generated as JS functions:

  var my_package_ClassName = function(...) {
    //constructor code
  };
  $hxClasses["my.package.ClassName"] = my_package_ClassName;
  my_package_ClassName.__name__ = "my.package.ClassName";
  my_package_ClassName.myStaticFunction = function() {...};
  my_package_ClassName.prototype = {
    ...
    ,myInstanceFunction: function() {...}
    ,__class__: my_package_ClassName
    ,__properties__: {...}
  };

Static functions calls are generated like this:

  my_package_ClassName.myStaticFunction()

When this function appears in a v8 stack trace, these
properties are present:

  type: Function
  function: my_package_ClassName.myStaticFunction
  method: myStaticFunction

This works as expected with the previous code.

Constructor function calls are generated like this:

  new my_package_ClassName()

When this function appears in a v8 stack trace, these
properties are present:

  type: null
  function: my_package_ClassName
  method: __class__

This will not work with the previous code, because the type is not
included in site.getFunctionName, and the function is the name of
the type. I'm not sure where else we might see a null type, but
we can determine this to be a constructor based on the method name
being __class__.

Instance function calls are generated like this:

  myClassInstance.myInstanceFunction()

When this function appears in a v8 stack trace, these
properties are present:

  type: my_package_ClassName
  function: myInstanceFunction
  method: myInstanceFunction

This also will not work with the previous code, because the type is not
included in site.getFunctionName.